topic 17 add kaminari
