+++ Gemfile
@@ -20,7 +20,7 @@ gem 'jbuilder', '~> 2.7'
 gem 'rails-i18n', '~> 6.0.0'
 gem 'aasm', '~> 4.11'
 gem 'ransack'
-
+gem 'kaminari'
 # Use Redis adapter to run Action Cable in production
 # gem 'redis', '~> 4.0'
 # Use Active Model has_secure_password
