+++ app/views/tasks/index.html.erb
@@ -39,6 +39,8 @@
   </tbody>
 </table>
 
+<%= paginate @tasks %>
+
