+++ config/locales/views/zh-TW.yml
@@ -1,4 +1,11 @@
 zh-TW:
+  views:
+    pagination:
+      first: "&laquo; 首頁"
+      last: "末頁 &raquo;"
+      next: "下一頁 &rsaquo;"
+      previous: "&lsaquo; 上一頁"
+      truncate: "&hellip;"
   tasks:
     show:
       title: "任務詳細"
